Headless, now available without UI.
I removed the hard dependency on the wx library. Since my plan is to make it run headless, it
doesn't make sense to depend on a UI library. If the library can be imported then it will still
import and work as it had but it won't crash if it doesn't exists now.

The obd_io file had an awful mess of 2, 3, 4, and 5 space indents. I've cleaned it all up to be
2 space indents only. I'm not sure if this was due to multiple parties working on the code
originally or was done to differentiate different types of blocks (e.g. class, def, if-else) but
it was a nightmare to work on in nano.