No point being lazy
I was a bit bored so I changed all the indentation to 2 spaces. Very therapeutic.

I also shortened the list of sensors I'm requesting in the headless entry point. The sensors I picked ended up causing issues when tested in a car, so hopefully this way I can narrow down the issues since I have an idea what the returns SHOULD be for those sensors