Created a Logger
I was using print statements in a few places for debugging and there's a chance I might miss one or two of them if I go back later to remove them.
So I created a Logger singleton class that checks for "--debug" as a command line argument for the obd_headless script which will enable the logger to print out the messages passed to it.
Otherwise it simply ignores any messages thrown at it.