Some syntactical changes
Fixed some syntax for my own sanity.
I introduced a lot of repetitive if statements checking if the wx module was importable which are now routed through a method.
Changed ```if self.port``` to ```if self.port is not None``` to be a little less ambiguous in case it caused issue